integrating ketcher,jsdraw,fileselect into core
