+++ package.dev.json
@@ -11,21 +11,20 @@
     "build-file-select": "ng build --prod file-select && cd dist/file-select && npm pack",
     "build-jsdraw-wrapper": "ng build --prod jsdraw-wrapper && cd dist/jsdraw-wrapper && npm pack",
     "build-ketcher-wrapper": "ng build --prod ketcher-wrapper && cd dist/ketcher-wrapper && npm pack",
-    "build-libraries": "npm run build-file-select && npm run build-jsdraw-wrapper && npm run build-ketcher-wrapper",
     "build-server": "ng run gsrs-client:server",
-    "start:gsrs:dev": "npm run build-libraries && npm run process-dojo && ng serve --aot=true --configuration=fda.dev --sourceMap=true",
-    "start:gsrs:local": "npm run build-libraries && npm run process-dojo && ng serve --configuration=fda.local",
-    "start:gsrs:pre-prod": "npm run build-libraries && npm run process-dojo && ng serve --aot=true --configuration=fda.pre-prod --sourceMap=true",
-    "start:gsrs:prod": "npm run build-libraries && npm run process-dojo && ng serve --aot=true --configuration=fda.prod --sourceMap=true",
-    "build:gsrs:pre-prod": "npm run build-libraries && npm run clear-libs-src && ng build --base-href=/ginas/app/beta/ --aot=true --configuration=fda.pre-prod --sourceMap=true && node process-index.js && cd lib && extract-zip dojo-custom-jsdraw.zip && cd .. && npm run copy-libs-dist && npm run build-server",
-    "build:gsrs:prod": "npm run build-libraries && npm run clear-libs-src && ng build --base-href=/ginas/app/ui/ --configuration=fda.prod && node process-index.js && cd lib && extract-zip dojo-custom-jsdraw.zip && cd .. && npm run copy-libs-dist && npm run build-server",
-    "start:fda:dev": "npm run build-libraries && npm run process-dojo && ng serve --aot=true --configuration=fda.dev --sourceMap=true",
-    "start:fda:local": "npm run build-libraries && npm run process-dojo && ng serve --configuration=fda.local",
-    "start:fda:pre-prod": "npm run build-libraries && npm run process-dojo && ng serve --aot=true --configuration=fda.pre-prod --sourceMap=true",
-    "start:fda:prod": "npm run build-libraries && npm run process-dojo && ng serve --aot=true --configuration=fda.prod --sourceMap=true",
-    "build:fda:pre-prod": "npm run build-libraries && npm run clear-libs-src && ng build --base-href=/ginas/app/beta/ --aot=true --configuration=fda.pre-prod --sourceMap=true && node process-index.js && cd lib && extract-zip dojo-custom-jsdraw.zip && cd .. && npm run copy-libs-dist && npm run build-server",
-    "build:fda:prod": "npm run build-libraries && npm run clear-libs-src && ng build --base-href=/ginas/app/ui/ --configuration=fda.prod && node process-index.js && cd lib && extract-zip dojo-custom-jsdraw.zip && cd .. && npm run copy-libs-dist && npm run build-server",
-     "test": "ng test",
+    "start:gsrs:dev": "npm run process-dojo && ng serve --aot=true --configuration=fda.dev --sourceMap=true",
+    "start:gsrs:local": "npm run process-dojo && ng serve --configuration=fda.local",
+    "start:gsrs:pre-prod": "npm run process-dojo && ng serve --aot=true --configuration=fda.pre-prod --sourceMap=true",
+    "start:gsrs:prod": "npm run process-dojo && ng serve --aot=true --configuration=fda.prod --sourceMap=true",
+    "build:gsrs:pre-prod": "npm run clear-libs-src && ng build --base-href=/ginas/app/beta/ --aot=true --configuration=fda.pre-prod --sourceMap=true && node process-index.js && cd lib && extract-zip dojo-custom-jsdraw.zip && cd .. && npm run copy-libs-dist && npm run build-server",
+    "build:gsrs:prod": "npm run clear-libs-src && ng build --base-href=/ginas/app/ui/ --configuration=fda.prod && node process-index.js && cd lib && extract-zip dojo-custom-jsdraw.zip && cd .. && npm run copy-libs-dist && npm run build-server",
+    "start:fda:dev": "npm run process-dojo && ng serve --aot=true --configuration=fda.dev --sourceMap=true",
+    "start:fda:local": "npm run process-dojo && ng serve --configuration=fda.local",
+    "start:fda:pre-prod": "npm run process-dojo && ng serve --aot=true --configuration=fda.pre-prod --sourceMap=true",
+    "start:fda:prod": "npm run process-dojo && ng serve --aot=true --configuration=fda.prod --sourceMap=true",
+    "build:fda:pre-prod": "npm run clear-libs-src && ng build --base-href=/ginas/app/beta/ --aot=true --configuration=fda.pre-prod --sourceMap=true && node process-index.js && cd lib && extract-zip dojo-custom-jsdraw.zip && cd .. && npm run copy-libs-dist && npm run build-server",
+    "build:fda:prod": "npm run clear-libs-src && ng build --base-href=/ginas/app/ui/ --configuration=fda.prod && node process-index.js && cd lib && extract-zip dojo-custom-jsdraw.zip && cd .. && npm run copy-libs-dist && npm run build-server",
+      "test": "ng test",
     "lint": "ng lint",
     "tslint": "tslint",
     "e2e": "ng e2e",
