--- projects/file-select/src/lib/file-select.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { FileSelectDirective } from './file-select.directive';
-
-@NgModule({
-  imports: [
-  ],
-  declarations: [FileSelectDirective],
-  exports: [FileSelectDirective]
-})
-export class FileSelectModule { }
