--- projects/file-select/src/lib/html-input-event.model.ts
@@ -1,3 +1,0 @@
-export interface HTMLInputEvent extends Event {
-    target: HTMLInputElement & EventTarget;
-}
