--- projects/jsdraw-wrapper/src/lib/jsdraw-wrapper.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { JsdrawWrapperComponent } from './jsdraw-wrapper.component';
-
-@NgModule({
-  imports: [
-  ],
-  declarations: [JsdrawWrapperComponent],
-  exports: [JsdrawWrapperComponent]
-})
-export class JsdrawWrapperModule { }
