--- projects/jsdraw-wrapper/src/lib/jsdraw.d.ts
@@ -1,6 +1,0 @@
-import { Injectable } from "@angular/core";
-
-@Injectable()
-export class JSDraw {
-    constructor(elementId: string);
-}
