--- projects/ketcher-wrapper/src/lib/ketcher-wrapper.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { KetcherWrapperComponent } from './ketcher-wrapper.component';
-
-@NgModule({
-  imports: [
-  ],
-  declarations: [KetcherWrapperComponent],
-  exports: [KetcherWrapperComponent]
-})
-export class KetcherWrapperModule { }
