--- projects/ketcher-wrapper/src/lib/ketcher-wrapper.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class KetcherWrapperService {
-
-  constructor() { }
-}
