+++ src/app/core/file-select/file-select.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { FileSelectDirective } from './file-select.directive';
+
+@NgModule({
+  imports: [
+  ],
+  declarations: [FileSelectDirective],
+  exports: [FileSelectDirective]
+})
+export class FileSelectModule { }
