+++ src/app/core/file-select/html-input-event.model.ts
@@ -1,0 +1,3 @@
+export interface HTMLInputEvent extends Event {
+    target: HTMLInputElement & EventTarget;
+}
