+++ src/app/core/structure-editor/jsdraw-wrapper/jsdraw-wrapper.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { JsdrawWrapperComponent } from './jsdraw-wrapper.component';
+
+@NgModule({
+  imports: [
+  ],
+  declarations: [JsdrawWrapperComponent],
+  exports: [JsdrawWrapperComponent]
+})
+export class JsdrawWrapperModule { }
