+++ src/app/core/structure-editor/jsdraw-wrapper/jsdraw.d.ts
@@ -1,0 +1,6 @@
+import { Injectable } from "@angular/core";
+
+@Injectable()
+export class JSDraw {
+    constructor(elementId: string);
+}
