+++ src/app/core/structure-editor/ketcher-wrapper/ketcher-wrapper.component.html
