+++ src/app/core/structure-editor/ketcher-wrapper/ketcher-wrapper.component.scss
@@ -1,0 +1,21 @@
+#ketcher-frame {
+    width: 100%;
+    height: 100%;
+    border: none;
+}
+
+::ng-deep button[title='3D Viewer'] {
+    display:none;
+  }
+  ::ng-deep  .LeftToolbar-module_group__0s41t.LeftToolbar-module_groupItem__OqQu0:has(button[data-testid=\'reaction-plus\']) {
+    display: none;
+  }
+  ::ng-deep .LeftToolbar-module_group__0s41t.LeftToolbar-module_groupItem__OqQu0:has(button[data-testid=\'shape-ellipse\']) {
+    display: none;
+  }
+  div#root{
+     border: 0.001em solid white;
+  }
+  div#root:focus-within {
+    border: 0.001em solid blue !important;
+  }
