+++ src/app/core/structure-editor/ketcher-wrapper/ketcher-wrapper.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { KetcherWrapperComponent } from './ketcher-wrapper.component';
+
+@NgModule({
+  imports: [
+  ],
+  declarations: [KetcherWrapperComponent],
+  exports: [KetcherWrapperComponent]
+})
+export class KetcherWrapperModule { }
