+++ src/app/core/structure-editor/ketcher-wrapper/ketcher-wrapper.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class KetcherWrapperService {
+
+  constructor() { }
+}
